mise a jout du projet

+++ src/main/java/com/quamtech/inventory_management/service/LocationService.java
@@ -1,6 +1,7 @@
 package com.quamtech.inventory_management.service;
 
 import com.quamtech.inventory_management.entite.Location;
+import com.quamtech.inventory_management.exception.InventoryException;
 import com.quamtech.inventory_management.payload.request.LocationRequest;
 import com.quamtech.inventory_management.repository.LocationRepository;
 import lombok.RequiredArgsConstructor;
@@ -27,7 +28,7 @@ public class LocationService {
                 .build();
         return locationRepository.save(createEmplacement1);
     }
-    public Location updateLocation(String id, LocationRequest locationRequest) {
+    public Location updateLocation(String id, LocationRequest locationRequest) throws InventoryException {
         Location existing = getLocationById(id);
         existing.setCode(locationRequest.getCode());
         existing.setWarehouseId(locationRequest.getWarehouseId());
@@ -41,20 +42,31 @@ public class LocationService {
         return locationRepository.save(existing);
     }
 
-    public Location getLocationById(String id) {
+    public Location getLocationById(String id) throws InventoryException {
         return locationRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Emplacement non trouvé avec l'ID: " + id));
+                .orElseThrow(() -> new InventoryException("Emplacement non trouvé avec l'ID: " + id));
     }
 
-    public List<Location> getAllLocations() {
+    public List<Location> getAllLocations() throws InventoryException {
+        List<Location> locationList=locationRepository.findAll();
+        if (locationList.isEmpty()){
+            throw new InventoryException("liste vide");
+        }
         return locationRepository.findAll();
     }
 
-    public List<Location> getLocationsByWarehouse(String warehouseId) {
-        return locationRepository.findByWarehouseIdAndActiveTrue(warehouseId);
+    public List<Location> getLocationsByWarehouse(String warehouseId) throws InventoryException {
+        List<Location> locationList=locationRepository.findByWarehouseIdAndActiveTrue(warehouseId);
+        if (locationList.isEmpty()){
+            throw new InventoryException("l'emplacement avec l'id"+warehouseId+"n'existe pas");
+        }
+        return locationList ;
     }
 
-    public void deleteLocation(String id) {
+    public void deleteLocation(String id) throws InventoryException {
+        if (!locationRepository.existsById(id)){
+            throw new InventoryException("cet id n'existe par");
+        }
         locationRepository.deleteById(id);
     }
